Added an example of bitmap writes
I've added a new example which assumes you have 4 of the 8x32 displays
arranged in a 16x64 display. It copies bitmaps out of a 16x16 font onto the
displays. The bit and pointer maths is all far too optimised for this to be
easy for beginners to follow. This may or may not be addressed in the future.