Explicitly set clock source to support HT1632C
The HT1632C revision changed the way master mode was set. There is no longer a specific command for setting the master mode, it just uses the clock source commands. This patch adds an explicit call to set the clock source to internal which should support both the HT1632 and the new HT1632C.